pyproject.toml
[tool.poetry]
name = "tsukuyomi"
version = "0.1.0"
description = "mettya　kashikoi"
authors = ["northfieldzz <northfield.t@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
discord-py = {extras = ["voice"], version = "^2.2.3"}
waitress = "^2.1.2"
requests = "^2.31.0"
setuptools = "^67.8.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"